+++ src/CanComment.php
@@ -1,52 +1,50 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: actuallymab
- * Date: 12.06.2016
- * Time: 02:13
- */
+declare(strict_types=1);
 
 namespace Actuallymab\LaravelComment;
 
-use Actuallymab\LaravelComment\Models\Comment;
+use Actuallymab\LaravelComment\Contracts\Commentable;
+use Illuminate\Database\Eloquent\Relations\MorphMany;
 
 trait CanComment
 {
-
-    /**
-     * @param $commentable
-     * @param string $commentText
-     * @param int $rate
-     * @return $this
-     */
-    public function comment($commentable, $commentText = '', $rate = 0)
+    public function comment(Commentable $commentable, string $commentText = '', int $rate = 0): self
     {
-        $comment = new Comment([
-            'comment'        => $commentText,
-            'rate'           => ($commentable->getCanBeRated()) ? $rate : null,
-            'approved'       => ($commentable->mustBeApproved() && ! $this->isAdmin()) ? false : true,
-            'commented_id'   => $this->id,
-            'commented_type' => $this->getMorphClass()
-        ]);
+        $commentModel = config('comment.model');
 
-        $commentable->comments()->save($comment);
+        $commentable->comments()->save(new $commentModel([
+            'comment'        => $commentText,
+            'rate'           => $commentable->canBeRated() ? $rate : null,
+            'approved'       => $commentable->mustBeApproved() && !$this->canCommentWithoutApprove() ? false : true,
+            'commented_id'   => $this->primaryId(),
+            'commented_type' => get_class(),
+        ]));
 
         return $this;
     }
 
-    /**
-     * @return bool
-     */
-    public function isAdmin()
+    public function canCommentWithoutApprove(): bool
     {
         return false;
     }
 
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function comments()
+    public function comments(): MorphMany
     {
-        return $this->morphMany(Comment::class, 'commented');
+        return $this->morphMany(config('comment.model'), 'commented');
+    }
+
+    public function hasCommentsOn(Commentable $commentable): bool
+    {
+        return $this->comments()
+            ->where([
+                'commentable_id'   => $commentable->primaryId(),
+                'commentable_type' => get_class($commentable),
+            ])
+            ->exists();
+    }
+
+    private function primaryId(): string
+    {
+        return (string)$this->getAttribute($this->primaryKey);
     }
 }