+++ tests/Models/User.php
@@ -1,29 +1,29 @@
 <?php
-/** actuallymab | 12.06.2016 - 14:16 */
-
+declare(strict_types=1);
 
 namespace Actuallymab\LaravelComment\Tests\Models;
 
 use Actuallymab\LaravelComment\CanComment;
 use Illuminate\Database\Eloquent\Model;
 
+/**
+ * @property string $name
+ * @property boolean $is_admin
+ */
 class User extends Model
 {
     use CanComment;
 
-    public $isAdmin = false;
+    protected $guarded = [];
 
-    protected $fillable = [
-        'name'
+    protected $casts = [
+        'is_admin' => 'boolean',
     ];
 
     public $timestamps = false;
 
-    /**
-     * @return bool
-     */
-    public function isAdmin()
+    public function canCommentWithoutApprove(): bool
     {
-        return $this->isAdmin;
+        return $this->is_admin;
     }
 }