+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All Notable changes to `Laravel Comment` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [1.0.0] - 2018-11-21
+
+### Added
+- Initial Release.
+- `Commentable` has been reimplemented as a contract.
+- `HasComments` trait has been implemented.
+- `comment.php` configuration file has been created.
+
+## [0.3.0] - 2017-01-31
+
+### Added
+- Laravel 5.4 Support.
+
 ## [0.2.1] - 2016-10-20
 
 ### Added
@@ -15,6 +28,4 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Laravel 5.3 Support.
 
 ## [0.1.0] - 2016-06-11
-
-### Added
-- Initial Release.
+- Initial Release